requirements.prod.txt
Django>=4.2.0,<5.0.0
requests>=2.32.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
whitenoise>=6.6.0
django-storages[google]>=1.14.2
google-cloud-secret-manager>=2.17.0
google-cloud-storage>=2.14.0